Red Goblin & Lives Update
Added the Red Goblin and player Lives. Two issues to resolve: When a
green goblin is caught and there are red or blue goblins on the screen,
they reset. This should not happen, they should last the preset time.
Second issue, the red goblin has to spawn away from the player for the
game to be fair.